+++ src/command_def.ml
@@ -248,7 +248,7 @@ let default =
 
 let run ~version ~build_info () =
   (* Store begin time *)
-  let start = Time.(now () |> to_float) in
+  let start = Time.now () in
 
 
   (* XXX Hack to allow to run 'oclaunch 5' or 'oclaunch' as before, i.e. do not
@@ -310,9 +310,10 @@ let run ~version ~build_info () =
       | Error -> Messages.warning "Error with lockfile"
   ));
 
-  (* Display total running time, to float is a number of secconds *)
-  Messages.debug Time.(now () |> to_float |> (-.) start |> ( *. ) (-1.)
-  |> sprintf "Runned during %f second(s)");
+  (* Display total running time, pretty printing is handled by Time module *)
+  Messages.debug Time.(diff (now ()) start
+      |> Span.to_string_hum (* Round the value, 3 digits *)
+      |> sprintf "Duration: %s");
 
   (* Reset display *)
   Messages.reset ();